fix(TODO-282): double to floatキャストエラーを修正
Mathf.RoundToIntに渡すdouble値をfloatにキャストしてコンパイルエラーを解決

+++ jp.iridescent.multitimelinerecorder/Editor/MultiTimelineRecorder.cs
@@ -865,8 +865,8 @@ namespace Unity.MultiTimelineRecorder
                                 string rangeText;
                                 if (showTimingInFrames)
                                 {
-                                    int startFrame = Mathf.RoundToInt(recordingRange.startTime * frameRate);
-                                    int endFrame = Mathf.RoundToInt(recordingRange.endTime * frameRate);
+                                    int startFrame = Mathf.RoundToInt((float)(recordingRange.startTime * frameRate));
+                                    int endFrame = Mathf.RoundToInt((float)(recordingRange.endTime * frameRate));
                                     rangeText = $"({startFrame}f-{endFrame}f)";
                                 }
                                 else
